+++ src/index.ts
@@ -50,7 +50,7 @@ export {
 
 export { replayIntegration, browserTracingIntegration, registerSpanErrorInstrumentation, logger } from '@sentry/browser';
 
-export { NativePauseAppHangTracking as pauseAppHangTracking, NativeResumeAppHangTracking as resumeAppHangTracking } from './wrapper';
+export { pauseAppHangTracking, resumeAppHangTracking } from './wrapper';
 
 export { SDK_NAME, SDK_VERSION } from './version';
 export type { CapacitorOptions } from './options';